refactor(allocator/pool): share `AllocatorGuard` between pool impls (#13621)
Refactor.

We have 2 versions of `AllocatorPool` (standard and fixed-size). Aim is to combine them into one. This is a step along that way.

Previously both versions of `AllocatorPool` had very similar `AllocatorGuard` implementations. Replace them with a single `AllocatorGuard` implementation, which both versions share.
